Materialize metadata definitions from folder
